fix: sketch 导出没有 images 目录

+++ index.js
@@ -44,7 +44,9 @@ module.exports = function (source, callback) {
             return;
         }
         // 复制图片到结果文件夹
-        fse.copySync('./output/images', './output/html/images');
+        if (fs.existsSync()) {
+            fse.copySync('./output/images', './output/html/images');
+        }
         // 复制模板资源文件夹
         fse.copySync('./template/assets', './output/html/assets');
         // 复制首页
